Add jenkins-client library for simpler communication with Jenkins.
Reformat existing code to new standard.
Add provided type for dependencies.

Previously, compile dependencies were not getting copied into the final
JAR. This would cause a ClassNotFoundException on the server. The
solution is to manually tell gradle to copy these dependencies. However,
the use of a new 'provided' type for dependencies is necessary to avoid
copying the compile dependencies that are known to be already present on
the server.

Change-Id: Ic96ba74119f50aaf50d2cf8a8604975bdefbcb76