Fix bugs in provisioning and templates. Change default branch regexes.
Some notes on the string operations applied to the regexes:

Due to the design of the Gerrit event stream, branches under "refs/heads/" are
referenced by their shortname, i.e. "develop" rather than "refs/heads/develop". All
other branches, however, are referenced by the full ref name, i.e.
"refs/sandbox/my-change" or "refs/feature/fix-all-bugs". This, of course, makes the
branch regexes hard to construct.

To conform to a standard, Gerrit-CI expects the user to enter the full ref name in the
TextBox. To match all branches under "refs/heads/", the user should enter
"refs/heads/.*" rather than ".*".

In order to make this compatible with the Gerrit trigger plugin, as described above,
we have to strip off the "refs/heads/" prefix of the user's regex. We replace this
with "(?!refs/)" which prevents the regex from matching any ref name that is outside
of "refs/heads/". This is best explained by an example: if the user intends to match
any branch inside "refs/heads/", they should enter "refs/heads/.*". After cutting off
the prefix, we are left with ".*". But this matches any ref name, and we only want it
to match the shortnames provided by the Gerrit events stream. So we add "(?!refs/)" to
ensure that the regex does not match any ref names beginning with "refs/".

Finally, for verify branch regexes we add "(?!refs/meta/)" to the beginning of the
string to prevent the regex from ever matching a branch inside of "refs/meta". We
also add the "^" and "$" characters to ensure that the regex is anchored.

Change-Id: I7903bec62cadf89b84dc1a2593d1fa63147d3f68