Single location for storing commands that ensure a clean start state
We store the verify/publish commands in Jenkins, but of course we add a
header that includes a bunch of commands to make sure we are in a clean
state.  That header was duplicated in both the verify and publish
templates, and we had a third (slightly modified) copy inside the code for
determining which part of the commands in Jenkins was just the standard
header.  Having three copies is problematic; store it in a single place.

Pair programmed with Elijah.