README.md: Move User Guide before installation and admin guide
People who are installing or administering a system running this plugin
will want to know everything in the installation, admin, and user guide
sections of this document.  However, users only care about the user guide,
and there will be a lot more of them than admins.  Since we are pointing
end-users at this documentation, let's put the User Guide first.